Add serial commands
